# Re-use label builder in promql aggregation