@codesome 2019-06-10
@jacksontj could you also add `BenchmarkEvaluations` that you used on this?