@jacksontj 2019-06-10
All the promql bench results:
```
benchmark                                                                                                             old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-8                                                                              39870          39264          -1.52%
BenchmarkRangeQuery/expr=a_one,steps=10-8                                                                             39204          39381          +0.45%
BenchmarkRangeQuery/expr=a_one,steps=100-8                                                                            51950          51465          -0.93%
BenchmarkRangeQuery/expr=a_one,steps=1000-8                                                                           150630         149374         -0.83%
BenchmarkRangeQuery/expr=a_ten,steps=1-8                                                                              139443         137220         -1.59%
BenchmarkRangeQuery/expr=a_ten,steps=10-8                                                                             142280         140581         -1.19%
BenchmarkRangeQuery/expr=a_ten,steps=100-8                                                                            260322         259950         -0.14%
BenchmarkRangeQuery/expr=a_ten,steps=1000-8                                                                           1230106        1225745        -0.35%
BenchmarkRangeQuery/expr=a_hundred,steps=1-8                                                                          1172953        1165467        -0.64%
BenchmarkRangeQuery/expr=a_hundred,steps=10-8                                                                         1209998        1200509        -0.78%
BenchmarkRangeQuery/expr=a_hundred,steps=100-8                                                                        2411998        2369116        -1.78%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-8                                                                       12033785       11662665       -3.08%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-8                                                                    53823          52308          -2.81%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-8                                                                   54662          56004          +2.46%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-8                                                                  91240          88437          -3.07%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-8                                                                 287896         283354         -1.58%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-8                                                                    164662         165886         +0.74%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-8                                                                   179764         177304         -1.37%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-8                                                                  413084         415430         +0.57%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-8                                                                 2489245        2372546        -4.69%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-8                                                                1338874        1327854        -0.82%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-8                                                               1475265        1459810        -1.05%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                                              4019102        3817567        -5.01%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-8                                                             26312639       23481218       -10.76%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-8                                                                2754490        2616156        -5.02%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-8                                                                26147286       25009536       -4.35%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-8                                                            265416052      249969932      -5.82%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-8                                                  1588782        1511588        -4.86%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-8                                                 2364456        2257370        -4.53%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-8                                                10316031       10111084       -1.99%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-8                                               92808502       89356432       -3.72%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-8                                                  11105076       10030729       -9.67%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-8                                                 18408799       17868157       -2.94%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-8                                                103763021      96642338       -6.86%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-8                                               939536966      889053030      -5.37%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-8                                              106911591      96813193       -9.45%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-8                                             181234067      175484187      -3.17%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8                                            1007580131     961770212      -4.55%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-8                                           9092937976     8845552222     -2.72%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-8                                                                 1183704        1173499        -0.86%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-8                                                                1663221        1632417        -1.85%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-8                                                               6409583        6353453        -0.88%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-8                                                              54448281       53640798       -1.48%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-8                                                                 9500849        9102089        -4.20%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-8                                                                14788819       13793484       -6.73%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-8                                                               65905785       61372729       -6.88%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-8                                                              556625572      542212428      -2.59%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-8                                                             96548813       88785119       -8.04%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-8                                                            141302217      136800958      -3.19%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-8                                                           632439271      612112543      -3.21%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-8                                                          5668862164     5394231035     -4.84%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-8                                                                    1206014        1141352        -5.36%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-8                                                                   1582995        1429440        -9.70%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-8                                                                  4904374        4371327        -10.87%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-8                                                                 40660803       33775760       -16.93%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-8                                                                    9439120        8681568        -8.03%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-8                                                                   12371773       11596796       -6.26%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-8                                                                  45317127       40949743       -9.64%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-8                                                                 371227061      335190451      -9.71%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-8                                                                89643179       83890754       -6.42%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-8                                                               124729787      115305286      -7.56%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                                              466495252      407510538      -12.64%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-8                                                             3793354356     3346217859     -11.79%
BenchmarkRangeQuery/expr=-a_one,steps=1-8                                                                             41055          40885          -0.41%
BenchmarkRangeQuery/expr=-a_one,steps=10-8                                                                            40935          41077          +0.35%
BenchmarkRangeQuery/expr=-a_one,steps=100-8                                                                           54764          53947          -1.49%
BenchmarkRangeQuery/expr=-a_one,steps=1000-8                                                                          156576         150860         -3.65%
BenchmarkRangeQuery/expr=-a_ten,steps=1-8                                                                             148972         147461         -1.01%
BenchmarkRangeQuery/expr=-a_ten,steps=10-8                                                                            150725         148821         -1.26%
BenchmarkRangeQuery/expr=-a_ten,steps=100-8                                                                           272398         276164         +1.38%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-8                                                                          1239508        1220917        -1.50%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-8                                                                         1229816        1227309        -0.20%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-8                                                                        1252719        1255659        +0.23%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-8                                                                       2444870        2461875        +0.70%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-8                                                                      12076157       11994669       -0.67%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-8                                                                      77616          75879          -2.24%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-8                                                                     80569          79870          -0.87%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-8                                                                    206863         213938         +3.42%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-8                                                                   1441209        1461738        +1.42%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-8                                                                      302163         304644         +0.82%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-8                                                                     419182         405529         -3.26%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-8                                                                    1668516        1640839        -1.66%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-8                                                                   13521175       13329646       -1.42%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-8                                                              2717623        2669855        -1.76%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-8                                                             3720680        3650263        -1.89%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-8                                                            16217046       15470934       -4.60%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-8                                                           140279429      132963266      -5.22%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-8                                                                  14561597       14213107       -2.39%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-8                                                                  139417654      137068903      -1.68%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          1398757265     1359811022     -2.78%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-8                                                     147615         144767         -1.93%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-8                                                    153665         148723         -3.22%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-8                                                   211065         207551         -1.66%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-8                                                  767609         744087         -3.06%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-8                                                     336075         319634         -4.89%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-8                                                    381004         370188         -2.84%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-8                                                   988421         983597         -0.49%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                  6659170        6470660        -2.83%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-8                                             2111042        2025791        -4.04%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-8                                            2570549        2513879        -2.20%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-8                                           8767671        8631770        -1.55%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                          69357003       68036403       -1.90%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-8                                                      151113         145823         -3.50%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-8                                                     158437         151197         -4.57%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-8                                                    234780         233093         -0.72%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-8                                                   1004372        1008930        +0.45%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-8                                                      339657         328149         -3.39%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-8                                                     413082         398599         -3.51%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-8                                                    1252133        1209207        -3.43%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                   9350641        9172774        -1.90%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-8                                              2192994        2073809        -5.43%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-8                                             2879561        2727372        -5.29%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-8                                            11175055       10721107       -4.06%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                           95534054       91188641       -4.55%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-8                                                  148980         147026         -1.31%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=10-8                                                 154793         153538         -0.81%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=100-8                                                233475         231647         -0.78%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-8                                               984296         996850         +1.28%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-8                                                  343857         323330         -5.97%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-8                                                 379304         366906         -3.27%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=100-8                                                987044         965643         -2.17%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-8                                               6579382        6471373        -1.64%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-8                                          2153135        2008807        -6.70%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=10-8                                         2534061        2470246        -2.52%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-8                                        8882158        8653796        -2.57%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                       68807870       68565490       -0.35%
…